Se realiza conexion myo y GUI falta prueba de escritorio y ver los datos
