+++ GUI_Adquisicion/GUI_adquisicion_funcional.py
@@ -27,7 +27,10 @@ from tkinter import *
 import sys
 
 import myo
-
+import subprocess
+import os
+from MYO_conexion import *
+import signal
 
 
 ###########################################################################
@@ -662,12 +665,16 @@ class FrameGesto1 (wx.Frame):
 
     def OnClickInicio(self, event):
         global i
+        global procesoEMG
         i = 0
+        procesoEMG = subprocess.Popen("python3 MYO_conexion.py", stdout=subprocess.PIPE, shell=True, preexec_fn=os.setsid)
         self.led.SetValue("0:00")
         self.OnTimer(None, e=10)
-        self.conexionMyo()
+        
+    
 
     def OnTimer(self, event, e):
+        global procesoEMG
         global i
         global c
         c = e
@@ -685,15 +692,8 @@ class FrameGesto1 (wx.Frame):
         else:
             print("stop")
             self.timer.Stop()
-            listener = EmgCollector(512)
-            Plot(listener).main(0)
+            os.killpg(os.getpgid(procesoEMG.pid), signal.SIGTERM)
 
-    def conexionMyo(self):
-        myo.init()
-        hub = myo.Hub()
-        listener = EmgCollector(512)
-        with hub.run_in_background(listener.on_event):
-            Plot(listener).main(1)
 
     def TimerGo(self, event):
         global s
@@ -720,64 +720,3 @@ class FrameGesto1 (wx.Frame):
         self.led.SetValue(t)
         self.OnTimer(None , c)
     
-### conexion Myo
-class EmgCollector(myo.DeviceListener):
-  """
-  Collects EMG data in a queue with *n* maximum number of elements.
-  """
-
-  def __init__(self, n):
-    self.n = n
-    self.lock = Lock()
-    self.emg_data_queue = deque(maxlen=n)
-
-  def get_emg_data(self):
-    with self.lock:
-      return list(self.emg_data_queue)
-
-  # myo.DeviceListener
-
-  def on_connected(self, event):
-    event.device.stream_emg(True)
-
-  def on_emg(self, event):
-    with self.lock:
-      self.emg_data_queue.append((event.timestamp, event.emg))
-
-class Plot(object):
-
-  def __init__(self, listener):
-    self.n = listener.n
-    self.listener = listener
-    self.fig = plt.figure()
-    self.axes = [self.fig.add_subplot('81' + str(i)) for i in range(1, 9)]
-    [(ax.set_ylim([-100, 100])) for ax in self.axes]
-    self.graphs = [ax.plot(np.arange(self.n), np.zeros(self.n))[0] for ax in self.axes]
-    plt.ion()
-
-  def update_plot(self):
-    # t_eeg = np.arange(0.0, 3.0, 0.01)
-    # data_eeg = np.sin(10 * np.pi * t_eeg)
-    # emg_data = self.listener.get_emg_data()
-    emg_data = self.listener.get_emg_data()
-    emg_data = np.array([x[1] for x in emg_data]).T
-    for g, data in zip(self.graphs, emg_data):
-      if len(data) < self.n:
-        # Fill the left side with zeroes.
-        data = np.concatenate([np.zeros(self.n - len(data)), data])
-      g.set_ydata(data)
-    plt.draw()
-
-  def main(self, closeGrap):
-    print("entro esto")
-    print(closeGrap)
-    print("entro esto")
-    if (closeGrap == 1):
-      self.update_plot()
-      plt.pause(1.0 / 30)
-    else:
-        print("entro esto final")
-        print(closeGrap)
-        matplotlib.pyplot.close("all")
-        # sys.exit()
-        return
